Device management migrated from using API Keys to JWT tokens
