+++ app/src/main/java/no/nordicsemi/android/ei/viewmodels/ProjectViewModel.kt
@@ -68,8 +68,10 @@ import no.nordicsemi.android.ei.model.Device
 import no.nordicsemi.android.ei.model.InferencingMessage.InferencingRequest
 import no.nordicsemi.android.ei.model.Message.Sample
 import no.nordicsemi.android.ei.model.Sensor
+import no.nordicsemi.android.ei.repository.DevicesRepository
 import no.nordicsemi.android.ei.repository.ProjectDataRepository
 import no.nordicsemi.android.ei.repository.ProjectRepository
+import no.nordicsemi.android.ei.repository.UserDataRepository
 import no.nordicsemi.android.ei.util.ZipPackage
 import no.nordicsemi.android.ei.util.guard
 import no.nordicsemi.android.ei.viewmodels.event.Event
@@ -84,6 +86,7 @@ class ProjectViewModel @Inject constructor(
     @ApplicationContext context: Context,
     private val userManager: UserManager,
     private val projectRepository: ProjectRepository,
+    private val devicesRepository: DevicesRepository,
     private val client: OkHttpClient,
     private val gson: Gson
 ) : AndroidViewModel(context as Application), FirmwareUpgradeCallback<FirmwareUpgradeManager.State> {
@@ -99,6 +102,9 @@ class ProjectViewModel @Inject constructor(
 
     private val projectManager: ProjectManager
         get() = userComponentEntryPoint.getProjectManager()
+
+    private val userDataManager: UserDataRepository
+        get() = userComponentEntryPoint.userDataRepository()
 
     private val projectDataRepository: ProjectDataRepository
         get() = EntryPoints
@@ -260,9 +266,9 @@ class ProjectViewModel @Inject constructor(
             }
         }
         viewModelScope.launch(handler) {
-            projectRepository.listDevices(
+            devicesRepository.listDevices(
+                token = userDataManager.token,
                 projectId = projectDataRepository.project.id,
-                keys = projectDataRepository.developmentKeys
             ).let { response ->
                 guard(response.success) {
                     throw Throwable(response.error)
@@ -646,8 +652,8 @@ class ProjectViewModel @Inject constructor(
                     .also { isDeviceRenameRequested = false }
             }
         }) {
-            projectRepository.renameDevice(
-                apiKey = keys.apiKey,
+            devicesRepository.renameDevice(
+                token = userDataManager.token,
                 projectId = project.id,
                 deviceId = device.deviceId,
                 name = name
@@ -672,8 +678,8 @@ class ProjectViewModel @Inject constructor(
                     .also { isDeviceRenameRequested = false }
             }
         }) {
-            projectRepository.deleteDevice(
-                apiKey = keys.apiKey,
+            devicesRepository.deleteDevice(
+                token = userDataManager.token,
                 projectId = project.id,
                 deviceId = device.deviceId
             ).let { response ->
